requirements.txt
groq==0.36.0
streamlit==1.37.0
python-dotenv
 
